+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ---
 
+#### 12.06.2018
+
+* [Rust] Code Royale - Mines have priority over Queen's health (CodinGame)
+
 #### 11.06.2018
 
 * [Rust] Code Royale - Single get structures fn (CodinGame)